Itty Witty Tiny Little Fix
Looking to change the variable target_transform to self.target_transform, it was a global otherwise. I went through this series for the first time about 8 months ago, and it took my coding way up to the next level and was a big part of my breaking into the industry so I wanted to thank you. I know this is a small change and the missile is really just an example, but everything about this engine is so tidy and this seemed it could be tightened up. I think there's a little critter in the GamePadMgr too, something about an uncapitalized 'i' - I'll hit that one too. Just thank you so much for making that series - it's so incredibly dense and pleasant at the same time.